Assorted cleanup (#26)
* Assorted cleanup; Improvements to README.md

Rename "Dockerfile.scanner.in.clam" to "Dockerfile".

Move the README.md from `src` to the top level directory so it will be
rendered by GitHub for the project page.

Remove the GPLv2 license blurbs in the source files and replace them
with the BSD 3-Clause license, since we decided to go with BSD 3-Clause.

Minor formatting fixes for the `--help` output. Fix one copy-paste issue
regarding the `--allmatch` option.

Tidy the README.md and fix some some inconsistencies. Significantly
expand on the installation documentation, particularly in regards to
Docker.

* Rename 'deepscan' to 'scan'; Set up pyproject.toml

I renamed the 'deepscan' command to 'scan', for simplicity.

I replaced 'requirements.txt' with 'pyproject.toml'.
This enables us to build and install a Python package and script
(CLI executable) to the PATH.

For a pleasant UX, I named the script "archive".
This way, the command goes from:
$ python3 /path/to/main.py deepscan /path/to/archive

To:
$ archive scan /path/to/archive

The command also reads well for the 'unpack' feature:
$ archive unpack /path/to/archive

I also updated the README.md install and usage instructions.

* Fix up pytest unit test suite

The previous restructuring caused problems with module imports.

* Review fix: remove references to requirements.txt

* Readme: privileged mode needed for docker containers